Improve format handling. Fix return value.
The previous version just use different suffixes to set file format.
Word actually uses the origin format (with different suffix) in this case.

In this new version, format is set via numerical format tag.
So all format issue is correctly handled.

The previous return value for success is 1. Now it is 0